UPDATE: refactor Data Table component to comply with project standards
